Struct schemamama_rusqlite::SqliteAdapter

pub struct SqliteAdapter<'a> {
    // some fields omitted
}

An adapter that allows its migrations to act upon PostgreSQL connection transactions.

Methods

impl<'a> SqliteAdapter<'a>

fn new(connection: &'a SqliteConnection) -> SqliteAdapter

Create a new migrator tied to a PostgreSQL connection.

fn setup_schema(&self)

Create the tables Schemamama requires to keep track of schema state. If the tables already exist, this function has no operation.